aisi_train/docker_files/requirements.txt
tqdm==4.61.2
torchvision==0.10.0
requests==2.25.1
torch==1.9.0
pycocotools==2.0.4
uvicorn==0.15.0
numpy==1.22.0
fastapi==0.65.2
Pillow==8.2.0
scikit_learn==0.24.2
python-multipart
